Disable logging to syslog
ntp's logging module defaults to logging to syslog and only duplicates
certain message types (such as debug) to the console. Explicitly disable
logging to syslog and enable logging (only to) the terminal.

BTW: There's also a global variable syslog_file to divert messages to a
log file. But that is disabled by default.

Signed-off-by: Michael Weiser <michael.weiser@gmx.de>